fix: correctness + robustness from line-by-line audit
Adversarially-found, manually RED-verified. Each fix mirrors an existing
blessed pattern; valid streams are unchanged.

- qdf_generic.go: UnmarshalT did not reset dec.colMaxLen on acquire. It shares
  the decoder pool with unmarshal / unmarshalQuery / SetInput, which all reset
  it because a generated columnar DecodeQDF that errors between
  ReadColStructHeader and ClearColMaxLen returns the decoder with a stale bound.
  Without the reset, UnmarshalT could inherit that bound and have colLenOK
  spuriously reject valid slice lengths. Reset it like every sibling path.
- colcodegen.go: WriteFloat32Column (codegen) always wrote raw Float32bits,
  ignoring OptCanonical, so it diverged from the reflect colKindFloat32 path
  (which normalizes -0.0 -> +0.0 and NaN -> one quiet NaN under OptCanonical).
  Honor OptCanonical so codegen and reflect produce identical canonical wire.
  Test: TestCodegenFloat32ColumnCanonical (RED-verified).
- decoder_skip.go: the tagPackDeltaFor skip computed bodyBits = (n64-1)*bitsPer
  BEFORE bounding n64, so a large varuint count overflowed uint64 and wrapped to
  a small body size that passed the bounds check and desynced the skip cursor.
  Bound (n64-1) by the remaining bytes via DIVISION first, like
  readPackedDeltaForHeader (the other skip codecs already do). Test:
  TestSkipDeltaForCountOverflow (RED-verified).
- reflect_encode.go: the probe-and-grow projection probeBytes*remaining could
  exceed int32 on a 32-bit build, wrapping the slices.Grow argument negative and
  panicking. Compute the product in int64 (two sites).

+++ reflect_encode.go
@@ -271,7 +271,10 @@ func encodeSlice(elem *typeDesc, stride uintptr, colPlan *columnarPlan) func(*En
 			// without forcing another doubling on a slight
 			// underestimate.
 			remaining := n - sliceProbeSize
-			projected := probeBytes * remaining / sliceProbeSize
+			// 64-bit intermediate: probeBytes*remaining can exceed int32 on a
+			// 32-bit build for a large slice, wrapping projected negative and
+			// panicking slices.Grow.
+			projected := int(int64(probeBytes) * int64(remaining) / int64(sliceProbeSize))
 			projected += projected >> 2
 			e.buf = slices.Grow(e.buf, projected)
 		}
@@ -1337,7 +1340,9 @@ func encodeSliceAny(e *Encoder, p unsafe.Pointer) error {
 	// Project the remaining size from the probe (+25% slack) and grow once,
 	// killing the doubling chain on large dynamic arrays.
 	if probeBytes := len(e.buf) - probeStart; probeBytes > 0 {
-		projected := probeBytes * (n - probe) / probe
+		// 64-bit intermediate: the product can exceed int32 on a 32-bit build for
+		// a large slice, wrapping projected negative and panicking slices.Grow.
+		projected := int(int64(probeBytes) * int64(n-probe) / int64(probe))
 		projected += projected >> 2
 		e.buf = slices.Grow(e.buf, projected)
 	}
